#59 digiwf-s3-integration-service refactor namespace
