+++ digiwf-integrations/digiwf-s3-integration/digiwf-s3-integration-service/src/main/java/io/muenchendigital/digiwf/s3/integration/MicroServiceApplication.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
+ * der Landeshauptstadt München, 2022
+ */
+package io.muenchendigital.digiwf.s3.integration;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+
+/**
+ * Application class for starting the micro-service.
+ */
+@SpringBootApplication
+@EnableJpaAuditing
+@EnableScheduling
+public class MicroServiceApplication {
+
+
+    public static void main(String[] args) {
+        SpringApplication.run(MicroServiceApplication.class, args);
+    }
+
+
+}
