+++ digiwf-integrations/digiwf-s3-integration/digiwf-s3-integration-service/src/main/java/io/muenchendigital/digiwf/s3/integration/configuration/NoSecurityConfiguration.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
+ * der Landeshauptstadt München, 2022
+ */
+package io.muenchendigital.digiwf.s3.integration.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
+
+
+@Configuration
+@Profile("no-security")
+@EnableResourceServer
+@EnableWebSecurity
+public class NoSecurityConfiguration extends ResourceServerConfigurerAdapter {
+
+    @Override
+    public void configure(ResourceServerSecurityConfigurer resources) {
+        resources.resourceId(null);
+    }
+
+    @Override
+    public void configure(HttpSecurity http) throws Exception {
+        http
+                .headers()
+                    .frameOptions()
+                    .disable()
+                .and().antMatcher("/**")
+                    .authorizeRequests()
+                    .anyRequest()
+                    .permitAll()
+                .and().csrf()
+                    .disable();
+    }
+
+}
