+++ digiwf-integrations/digiwf-s3-integration/digiwf-s3-integration-service/src/test/java/io/muenchendigital/digiwf/s3/integration/TestConstants.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
+ * der Landeshauptstadt München, 2022
+ */
+package io.muenchendigital.digiwf.s3.integration;
+
+import lombok.AccessLevel;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+import org.springframework.hateoas.RepresentationModel;
+
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class TestConstants {
+
+    public static final String SPRING_TEST_PROFILE = "test";
+
+    public static final String SPRING_NO_SECURITY_PROFILE = "no-security";
+
+}
